Add a method to check if the icon is valid
This will add a method to check if the icon part of the SIDC is valid
or not #24, if more validation is needed please add a new issue with
details of how the validation should work.
If it wasn’t possible to find an icon, an upside down question mark
will still be drawn as the icon.